images_microservice/requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
fastapi==0.115.4
fastapi-cache2==0.2.2
flatbuffers==24.3.25
gunicorn==23.0.0
h11==0.14.0
humanfriendly==10.0
idna==3.10
imageio==2.36.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
lazy_loader==0.4
llvmlite==0.43.0
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
packaging==24.2
pendulum==3.0.0
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.48
protobuf==5.28.3
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
PyMatting==1.1.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
redis==5.2.0
referencing==0.35.1
rembg==2.0.59
requests==2.32.3
rpds-py==0.21.0
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
sympy==1.13.3
tifffile==2024.9.20
time-machine==2.16.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
vine==5.1.0
wcwidth==0.2.13
